made apt reporting aware of phased update and general fixes/cleanup

+++ setup.sh
@@ -2,9 +2,15 @@
 
 #NOTE this is broken in this public version since I stripped a bunch of private info needed to make this work. I'll fix it eventually
 
+if ! grep -Eq "ID(_LIKE)?=debian" </etc/os-release ; then
+	printf "ERROR: only debian-based systems are supported, yours (%s) doesn't seem like one.\n" "$(grep -E "^ID(_LIKE)?" < /etc/os-release |tr '\n' ','|rev|cut -c2-|rev)"
+	printf "You may still use these tools, but none of it will be set up automatically.\nCheck this file to see what would have been done.\n"
+	exit 1
+fi
+
 { WSL_VERSION="$(/mnt/c/Windows/System32/wsl.exe -l -v | iconv -f unicode|grep -E "\b$WSL_DISTRO_NAME\s+Running" | awk '{print $NF}' | cut -c-1)"; } 2> /dev/null || true
 if [ "$WSL_VERSION" -eq 2 ] ; then
-	printf "You are running %s as WSL 2. are you REALLY sure you wouldn't rather have WSL 1?\nWSL1 is way way faster if you also need to access files from windows which is the entire point of WSL, else you'd just use a VM or a seperate Linux system\nIf so select no here, exit wsl and in Powershell run the followung commands\n\n\twsl --set-default-version 1\n\twsl --set-version %s 1\n\nDo you really want to continue in WSL2? if so enter 'yes' " "$WSL_DISTRO_NAME" "$WSL_DISTRO_NAME"
+	printf "You are running %s as WSL 2. are you REALLY sure you wouldn't rather have WSL 1?\nWSL1 is way way faster if you also need to access files from windows which is the entire point of WSL, else you'd just use a VM or a seperate Linux system\nIf so select no here, exit wsl and in Powershell run the following commands\n\n\twsl --set-default-version 1\n\twsl --set-version %s 1\n\nDo you really want to continue in WSL2? if so enter 'yes' " "$WSL_DISTRO_NAME" "$WSL_DISTRO_NAME"
 	read -r versionconfirm
 	case $versionconfirm in
 		yes )
@@ -75,11 +81,16 @@ zsh -ic "$cmd; if [ \$? -eq 0 ] ; then echo \"done\"; exit ; else printf \"somet
 zsh -ic "UpdateZSH;code 2>/dev/null;exit"
 #this is a bit unsatisfying...
 #if vs code isn't installed, install it
-#if ! which fuck | grep -vq 'not found' ; then echo "INSTALL" ; else echo "DONT" ; fi
+#if ! which inexistantcommand | grep -vq 'not found' ; then echo "INSTALL" ; else echo "DONT" ; fi
 if ! which code | grep -vq 'not found' ; then
-	curl -L -o vsc.deb "https://code.visualstudio.com/sha/download?build=stable&os=linux-deb-x64"
-	sudo dpkg -i vsc.deb
-	rm vsc.dpkg
+	if [ "$(uname -m)" != "x86_64" ]; then
+		echo "since you aren't using x86_64 VS Code will NOT be Auto-Installed"
+	else
+		echo "getting/installing vs code" ;
+		curl -L -o vsc.deb "https://code.visualstudio.com/sha/download?build=stable&os=linux-deb-x64"
+		sudo dpkg -i vsc.deb
+		rm vsc.dpkg
+	fi
 else
         echo "'code' is already installed -> skipping"
 fi
